Added a base Result class
