+++ src/DeleteResult.php
@@ -2,28 +2,13 @@
 
 namespace MongoDB;
 
-use MongoDB\Driver\WriteResult;
 use MongoDB\Exception\BadMethodCallException;
 
 /**
  * Result class for a delete operation.
  */
-class DeleteResult
+class DeleteResult extends Result
 {
-    private $writeResult;
-    private $isAcknowledged;
-
-    /**
-     * Constructor.
-     *
-     * @param WriteResult $writeResult
-     */
-    public function __construct(WriteResult $writeResult)
-    {
-        $this->writeResult = $writeResult;
-        $this->isAcknowledged = $writeResult->isAcknowledged();
-    }
-
     /**
      * Return the number of documents that were deleted.
      *
@@ -40,18 +25,5 @@ class DeleteResult
         }
 
         throw BadMethodCallException::unacknowledgedWriteResultAccess(__METHOD__);
-    }
-
-    /**
-     * Return whether this delete was acknowledged by the server.
-     *
-     * If the delete was not acknowledged, other fields from the WriteResult
-     * (e.g. deletedCount) will be undefined.
-     *
-     * @return boolean
-     */
-    public function isAcknowledged()
-    {
-        return $this->isAcknowledged;
     }
 }
